+++ src/buffer/user_context.rs
@@ -20,23 +20,31 @@ impl Entry {
 pub enum Message {
     Whois(User),
     Query(User),
+    SingleClick(User),
 }
 
 #[derive(Debug, Clone)]
 pub enum Event {
     SendWhois(User),
     OpenQuery(User),
+    SingleClick(User),
 }
 
 pub fn update(message: Message) -> Event {
     match message {
         Message::Whois(user) => Event::SendWhois(user),
         Message::Query(user) => Event::OpenQuery(user),
+        Message::SingleClick(user) => Event::SingleClick(user),
     }
 }
 
 pub fn view<'a>(content: impl Into<Element<'a, Message>>, user: User) -> Element<'a, Message> {
     let entries = Entry::list();
+
+    let content = button(content)
+        .padding(0)
+        .style(theme::Button::Bare)
+        .on_press(Message::SingleClick(user.clone()));
 
     context_menu(content, entries, move |entry, length| {
         let (content, message) = match entry {