+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Added:
 - Configurable keyboard shortcuts for common actions, such as changing buffer focus, maximize / restoring buffer size, 
   cycling channels in the buffer and more! A new `keys` section has been added to the config file, reference the 
   [wiki](https://github.com/squidowl/halloy/wiki/Keyboard-shortcuts) for more details.
+- Single clicking on a user will insert nickname to input
 
 Fixed: